Update training configuration and dependencies
- Modified `requirements.txt` to use flexible versions of PyTorch and related libraries.
- Adjusted training parameters in `train.yaml`, reducing epochs from 20 to 8 and batch sizes from 32 and 64 to 8.
- Enhanced `train_gnn.py` with gradient accumulation and mixed precision training support.
- Fixed data type handling in `graph_models.py` for message passing layer.

+++ src/train/graph_models.py
@@ -68,7 +68,8 @@ class BipartiteMessagePassingLayer(nn.Module):
         # Messages from attributes to objects.
         attr_msgs = flat_attrs.index_select(0, src_attr)  # [E, Din_attr]
         attr_msgs = self.attr_to_obj(attr_msgs)  # [E, Din]
-        w = edge_weight.view(-1, 1)
+        attr_msgs = attr_msgs.to(flat_objects.dtype)
+        w = edge_weight.view(-1, 1).to(flat_objects.dtype)
         weighted_msgs = attr_msgs * w
 
         # Aggregate messages per object index.
